philately/requirements.txt
asgiref==3.3.1
boto3==1.17.54
botocore==1.20.54
Django==3.1.7
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-ses==2.0.0
future==0.18.2
jmespath==0.10.0
mysqlclient==2.0.3
Pillow==8.2.0
pkg-resources==0.0.0
python-dateutil==2.8.1
python-http-client==3.3.2
pytz==2021.1
s3transfer==0.4.0
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
